feat: TS integration, modular structure
