+++ rustfmt.toml
@@ -1,0 +1,2 @@
+unstable_features = true
+merge_imports = true
