+++ src/keys/keypair.rs
@@ -4,6 +4,8 @@ use napi_derive::js_function;
 use rand::rngs::OsRng;
 use sn_api::Keypair;
 
+use crate::util;
+
 #[js_function(0)]
 fn constructor(ctx: CallContext) -> Result<JsUndefined> {
     // No wrapping here, that will be done by static methods or other objects.
@@ -12,7 +14,7 @@ fn constructor(ctx: CallContext) -> Result<JsUndefined> {
 
 #[js_function(0)]
 fn new_ed25519(ctx: CallContext) -> Result<JsObject> {
-    let mut kp_js = crate::get_constructor(ctx.env, "Keypair")?.new(&[] as &[JsNull])?;
+    let mut kp_js = util::get_constructor(ctx.env, "Keypair")?.new(&[] as &[JsNull])?;
 
     let mut rng = OsRng;
     let kp = Keypair::new_ed25519(&mut rng);
@@ -28,7 +30,7 @@ fn secret_key(ctx: CallContext) -> Result<JsObject> {
 
     let sk = kp.secret_key().map_err(|e| Error::from_reason(format!("{:?}", e)))?;
 
-    let mut sk_js = crate::get_constructor(ctx.env, "SecretKey")?.new(&[] as &[JsNull])?;
+    let mut sk_js = util::get_constructor(ctx.env, "SecretKey")?.new(&[] as &[JsNull])?;
     ctx.env.wrap(&mut sk_js, sk)?;
 
     Ok(sk_js)
