+++ tsconfig.json
@@ -1,0 +1,10 @@
+{
+    "compilerOptions": {
+        "lib": [
+            "es2020"
+        ],
+        "module": "commonjs",
+        "target": "ES2020",
+        "strict": true,
+    }
+}
